+++ packages/instant-apps-components/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.0.0-beta.300
+
+### instant-apps-filter-list
+
+- add @Method to update t9n for filterLayerExpressions
+
 ## v1.0.0-beta.299
 
 ### instant-apps-scoreboard